+++ crates/cast/tests/e2e/account/add.rs
@@ -59,9 +59,9 @@ pub async fn test_happy_case() {
 #[tokio::test]
 pub async fn test_happy_case_add_profile() {
     let current_dir = Utf8PathBuf::from(duplicate_directory_with_salt(
-        CONTRACTS_DIR.to_string() + "/v1/map",
+        CONTRACTS_DIR.to_string() + "/map",
         "put",
-        "3",
+        "30",
     ));
     let accounts_file = "./accounts.json";