+++ crates/forge/src/lib.rs
@@ -9,7 +9,6 @@ use camino::{Utf8Path, Utf8PathBuf};
 
 use futures::StreamExt;
 use running::blocking_run_from_test;
-use serde::Deserialize;
 use tokio::sync::mpsc::{channel, Sender};
 
 use std::sync::Arc;
@@ -23,6 +22,7 @@ use cairo_lang_sierra::program::Function;
 use cairo_lang_sierra_to_casm::metadata::MetadataComputationConfig;
 use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
 use futures::stream::FuturesUnordered;
+use itertools::Itertools;
 
 use once_cell::sync::Lazy;
 use rand::{thread_rng, RngCore};
@@ -34,9 +34,8 @@ use crate::scarb::{ForgeConfig, ForkTarget, StarknetContractArtifacts};
 // pub use crate::collecting::CrateLocation;
 pub use crate::test_crate_summary::TestCrateSummary;
 
-use crate::collecting::{
-    collect_test_compilation_targets, compile_tests, filter_tests_from_crates, CompiledTestCrate,
-};
+use crate::collecting::{collect_test_compilation_targets, compile_tests, CompiledTestCrate};
+use crate::test_filter::TestsFilter;
 use test_collector::{FuzzerConfig, LinkedLibrary, TestCase};
 
 pub mod pretty_printing;
@@ -48,8 +47,10 @@ mod fuzzer;
 mod running;
 mod test_crate_summary;
 mod test_execution_syscall_handler;
+mod test_filter;
 
 const FUZZER_RUNS_DEFAULT: u32 = 256;
+pub const CACHE_DIR: &str = ".snfoundry_cache";
 
 static BUILTINS: Lazy<Vec<&str>> = Lazy::new(|| {
     vec![
@@ -65,12 +66,11 @@ static BUILTINS: Lazy<Vec<&str>> = Lazy::new(|| {
 });
 
 /// Configuration of the test runner
-#[derive(Deserialize, Debug, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub struct RunnerConfig {
     workspace_root: Utf8PathBuf,
-    test_name_filter: Option<String>,
-    exact_match: bool,
     exit_first: bool,
+    tests_filter: TestsFilter,
     fork_targets: Vec<ForkTarget>,
     fuzzer_runs: u32,
     fuzzer_seed: u64,
@@ -85,20 +85,20 @@ impl RunnerConfig {
     /// * `exact_match` - Should test names match the `test_name_filter` exactly
     /// * `exit_first` - Should runner exit after first failed test
     #[must_use]
-    #[allow(clippy::too_many_arguments)]
+    #[allow(clippy::too_many_arguments, clippy::fn_params_excessive_bools)]
     pub fn new(
         workspace_root: Utf8PathBuf,
         test_name_filter: Option<String>,
         exact_match: bool,
         exit_first: bool,
+        only_ignored: bool,
+        include_ignored: bool,
         fuzzer_runs: Option<u32>,
         fuzzer_seed: Option<u64>,
         forge_config_from_scarb: &ForgeConfig,
     ) -> Self {
         Self {
             workspace_root,
-            test_name_filter,
-            exact_match,
             exit_first: forge_config_from_scarb.exit_first || exit_first,
             fork_targets: forge_config_from_scarb.fork.clone(),
             fuzzer_runs: fuzzer_runs
@@ -107,6 +107,12 @@ impl RunnerConfig {
             fuzzer_seed: fuzzer_seed
                 .or(forge_config_from_scarb.fuzzer_seed)
                 .unwrap_or_else(|| thread_rng().next_u64()),
+            tests_filter: TestsFilter::from_flags(
+                test_name_filter,
+                exact_match,
+                only_ignored,
+                include_ignored,
+            ),
         }
     }
 }
@@ -219,7 +225,10 @@ pub async fn run(
         .map(|ct| ct.ensure_lib_file_exists(&temp_dir))
         .collect::<Result<_>>()?;
     let tests = compile_tests(&compilation_targets, &runner_params)?;
-    let tests = filter_tests_from_crates(tests, &runner_config);
+    let tests = tests
+        .into_iter()
+        .map(|tc| runner_config.tests_filter.filter_tests(tc))
+        .collect_vec();
 
     try_close_tmp_dir(temp_dir)?;
 
@@ -293,9 +302,19 @@ async fn run_tests_from_crate(
     let (send_shut_down, mut rec_shut_down) = channel(1);
 
     for case in test_cases.iter() {
-        let case_name = case.name.as_str();
+        let case_name = case.name.clone();
 
-        let function = runner.find_function(case_name)?;
+        if !runner_config
+            .tests_filter
+            .should_be_run_based_on_ignored(case)
+        {
+            tasks.push(tokio::task::spawn(async {
+                Ok(TestCaseSummary::Ignored { name: case_name })
+            }));
+            continue;
+        };
+
+        let function = runner.find_function(&case_name)?;
         let args = function_args(function, &BUILTINS);
 
         let case = Arc::new(case.clone());
@@ -315,13 +334,20 @@ async fn run_tests_from_crate(
     }
 
     let mut results = vec![];
+    let mut interrupted = false;
 
     while let Some(task) = tasks.next().await {
         let result = task??;
+        match result {
+            // Because tests are executed parallel is possible to receive
+            // Ok(TestCaseSummary::Interrupted) before Err
+            TestCaseSummary::Interrupted {} => interrupted = true,
+            result => {
+                pretty_printing::print_test_result(&result);
 
-        pretty_printing::print_test_result(&result);
-
-        results.push(result);
+                results.push(result);
+            }
+        }
     }
 
     rec.close();
@@ -329,6 +355,11 @@ async fn run_tests_from_crate(
     // Waiting for things to finish shutting down
     drop(send_shut_down);
     let _ = rec_shut_down.recv().await;
+
+    // This Panic should never occur.
+    // If TestCaseSummary::Interrupted is returned by a test,
+    // this implies that there should be another test than returned an Err.
+    assert!(!interrupted, "Tests were interrupted");
 
     let contained_fuzzed_tests = results.iter().any(|summary| summary.runs().is_some());
     Ok(TestCrateSummary {
@@ -393,7 +424,7 @@ fn run_single_test(
             () = cancellation_tokens.error.cancelled() => {
                 // Stop executing all tests because
                 // one of a test returns Err
-                Ok(TestCaseSummary::InterruptedByError {  })
+                Ok(TestCaseSummary::Interrupted{  })
             },
 
             result = blocking_run_from_test(vec![], case.clone(),runner,  runner_config.clone(), runner_params.clone(), send.clone(), send_shut_down.clone() ) => {
@@ -426,7 +457,7 @@ fn run_with_fuzzing(
     send_shut_down: Sender<()>,
 ) -> JoinHandle<Result<TestCaseSummary>> {
     tokio::task::spawn(async move {
-        let token = CancellationToken::new();
+        let cancellation_fuzzing_token = CancellationToken::new();
         let (send, mut rec) = channel(1);
         let args = args
             .iter()
@@ -459,7 +490,7 @@ fn run_with_fuzzing(
                 runner_config.clone(),
                 runner_params.clone(),
                 cancellation_tokens.clone(),
-                token.clone(),
+                cancellation_fuzzing_token.clone(),
                 send.clone(),
                 send_shut_down.clone(),
             ));
@@ -475,7 +506,11 @@ fn run_with_fuzzing(
             final_result = Some(result.clone());
 
             match result {
-                TestCaseSummary::Failed { .. } | TestCaseSummary::InterruptedByError {} => {
+                TestCaseSummary::Failed { .. } => {
+                    cancellation_fuzzing_token.cancel();
+                    break;
+                }
+                TestCaseSummary::Interrupted {} => {
                     break;
                 }
                 _ => (),
@@ -521,7 +556,7 @@ fn run_fuzzing_subtest(
             () = cancellation_tokens.error.cancelled() => {
                 // Stop executing all tests because
                 // one of a test returns Err
-                Ok(TestCaseSummary::InterruptedByError {  })
+                Ok(TestCaseSummary::Interrupted{  })
             },
             () = cancellation_tokens.exit_first.cancelled() => {
                 // Stop executing all tests because flag --exit-first'
@@ -531,7 +566,7 @@ fn run_fuzzing_subtest(
             () = cancellation_fuzzing_token.cancelled() => {
                 // Stop executing all single fuzzing tests
                 // because one of fuzzing test has been FAIL
-                Ok(TestCaseSummary::SkippedFuzzing {})
+                Ok(TestCaseSummary::Interrupted {  })
 
             },
            result = blocking_run_from_test(
@@ -548,8 +583,6 @@ fn run_fuzzing_subtest(
                         if let TestCaseSummary::Failed { .. } = &result {
                             if runner_config.exit_first {
                                 cancellation_tokens.exit_first.cancel();
-                            } else {
-                                cancellation_fuzzing_token.cancel();
                             }
                         }
                         Ok(result)
@@ -581,6 +614,7 @@ fn function_args<'a>(function: &'a Function, builtins: &[&str]) -> Vec<&'a Concr
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_filter::{IgnoredFilter, NameFilter};
 
     #[test]
     fn fuzzer_default_seed() {
@@ -590,6 +624,8 @@ mod tests {
             None,
             false,
             false,
+            false,
+            false,
             None,
             None,
             &Default::default(),
@@ -597,6 +633,8 @@ mod tests {
         let config2 = RunnerConfig::new(
             workspace_root,
             None,
+            false,
+            false,
             false,
             false,
             None,
@@ -617,6 +655,8 @@ mod tests {
             None,
             false,
             false,
+            false,
+            false,
             None,
             None,
             &Default::default(),
@@ -625,9 +665,11 @@ mod tests {
             config,
             RunnerConfig {
                 workspace_root,
-                test_name_filter: None,
-                exact_match: false,
                 exit_first: false,
+                tests_filter: TestsFilter {
+                    name_filter: NameFilter::All,
+                    ignored_filter: IgnoredFilter::NotIgnored
+                },
                 fork_targets: vec![],
                 fuzzer_runs: FUZZER_RUNS_DEFAULT,
                 fuzzer_seed: config.fuzzer_seed,
@@ -647,8 +689,10 @@ mod tests {
 
         let config = RunnerConfig::new(
             workspace_root.clone(),
-            None,
+            Some("test".to_string()),
             false,
+            false,
+            true,
             false,
             None,
             None,
@@ -658,10 +702,12 @@ mod tests {
             config,
             RunnerConfig {
                 workspace_root,
-                test_name_filter: None,
-                exact_match: false,
                 exit_first: true,
                 fork_targets: vec![],
+                tests_filter: TestsFilter {
+                    name_filter: NameFilter::Match("test".to_string()),
+                    ignored_filter: IgnoredFilter::Ignored
+                },
                 fuzzer_runs: 1234,
                 fuzzer_seed: 500,
             }
@@ -680,7 +726,9 @@ mod tests {
         };
         let config = RunnerConfig::new(
             workspace_root.clone(),
-            None,
+            Some("abc".to_string()),
+            true,
+            true,
             false,
             true,
             Some(100),
@@ -691,13 +739,47 @@ mod tests {
             config,
             RunnerConfig {
                 workspace_root,
-                test_name_filter: None,
-                exact_match: false,
                 exit_first: true,
+                tests_filter: TestsFilter {
+                    name_filter: NameFilter::ExactMatch("abc".to_string()),
+                    ignored_filter: IgnoredFilter::All
+                },
                 fork_targets: vec![],
                 fuzzer_runs: 100,
                 fuzzer_seed: 32,
             }
+        );
+    }
+
+    #[test]
+    #[should_panic]
+    fn only_ignored_and_include_ignored_both_true() {
+        let _ = RunnerConfig::new(
+            Default::default(),
+            None,
+            false,
+            false,
+            true,
+            true,
+            None,
+            None,
+            &Default::default(),
+        );
+    }
+
+    #[test]
+    #[should_panic]
+    fn exact_match_true_without_test_filter_name() {
+        let _ = RunnerConfig::new(
+            Default::default(),
+            None,
+            true,
+            false,
+            true,
+            true,
+            None,
+            None,
+            &Default::default(),
         );
     }
 }